XD-1930: Fix java8 compile warning
ContainerConfiguration is now a (still default visibility & top level)
non-auxiliary class.
